feat(001): fix mistakes, fix structure, refact
